third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-017-expected.txt


FAIL CSS Test: @viewport constrained - Large non-auto height not affecting width. assert_equals: expected 10000 but got 600
